pyproject.toml
[project]
name = "otel-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opentelemetry-api>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "pytest>=8.4.1",
    "ruff>=0.12.10",
]

[tool.pytest.ini_options]
python_files = "*_test.py"
